implemented no_rt mode (-no_rt) fixed fullscreen bug (+more minor ones) reduced compiler warnings optimized mesh generation

+++ liblava_rtt_extension/src/rt_pipeline.cpp
@@ -161,7 +161,7 @@ bool raytracing_pipeline::setup() {
     std::vector<uint8_t> handles(handle_size * groups.size());
 
     if (!check(device->call().vkGetRayTracingShaderGroupHandlesKHR(
-            device->get(), get(), 0, groups.size(), handles.size(), handles.data()))) {
+            device->get(), get(), 0, groups.size(), uint32_t(handles.size()), handles.data()))) {
         return false;
     }
